test: fold PostHog cases into analytics
- Move signal, no-op, shutdown, and before_send cases
- Remove the orphan posthog suite file
